SearchImage_App/requirements.txt
streamlit
sentence-transformers
faiss-cpu
pillow
numpy